+++ routes/index.js
@@ -9,12 +9,12 @@ var Q = require('q');
 var ensureLoggedIn = require('./middleware/ensureLoggedIn');
 var ensureLoggedOut = require('./middleware/ensureLoggedOut');
 
-var signup = require('./auth/signup');
+var request = require('./auth/request');
 var login = require('./auth/login');
 var admin = require('./auth/admin');
 
 router.use('/login/', login);
-router.use('/signup/', signup);
+router.use('/request/', request);
 router.use('/admin/', admin);
 
 /* GET index page. */
@@ -25,16 +25,19 @@ router.get('/library/', ensureLoggedIn, function(req, res, next) {
   // Retrieve the user from the session.
   var user = req.session.user;
   // Render the library page.
-  res.render('library', { user: user, client_id: config.auth.client_id});
+  res.render('library', {
+    user: user,
+    client_id: config.auth.client_id
+  });
 });
 
 /*GET logout page. */
-router.get('/logout/', ensureLoggedIn, function(req, res, next){
+router.get('/logout/', ensureLoggedIn, function(req, res, next) {
   // Destroy the session.
-  req.session.destroy(function(err){
-    console.log(err);
-  })
-  // Redirect to the home page.
+  req.session.destroy(function(err) {
+      console.log(err);
+    })
+    // Redirect to the home page.
   res.redirect('/login/');
 })
