+++ views/login.hbs
@@ -24,7 +24,7 @@
 
 											<button type="submit" value="Submit" class="btn btn-success btn-block">Submit</button>
 											<hr>
-											<a href="javascript:loadSignup();" class="btn btn-success btn-block">Signup</a>
+											<a href="/request/" class="btn btn-success btn-block">Request Account</a>
 								</form>
 							</div>
 					</div>
@@ -33,4 +33,4 @@
 </div>
 
 <script src="/bower_components/jquery/dist/jquery.min.js"></script>
-<script src="/js/login-signup.js"></script>
+<script src="/js/login-request.js"></script>
